+++ lib/rails/generators/fortitude/scaffold/scaffold_generator.rb
@@ -12,15 +12,6 @@ module Fortitude
         end
       end
 
-      hook_for :form_builder, :as => :scaffold
-
-      def copy_form_file
-        if options[:form_builder].nil?
-          filename = filename_with_extensions("_form")
-          template "_form.html.rb", File.join("app/views", controller_file_path, filename)
-        end
-      end
-
       def create_view_base
         generate "fortitude:base_view"
       end